Refactor plot type names and improve axis handling
Updated plot type names in plot_types.py for consistency and removed duplicates. Changed import paths for PlotPane to use monstim_gui.plotting. Renamed plot_reflexCurves to plot_averageReflexCurves in dataset and experiment plotters for clarity. Improved y-axis handling in session_plotter_pyqtgraph.py by unlinking y-axes before enabling auto-range, preventing unintended axis linking when plotting multiple channels. Added error handling for zero channels in base_plotter_pyqtgraph.py.
